Move "create" command to ant plugin. Add sdk:plugin:info command. Add newer version check and update to sdk:plugin:list.

+++ driver/src/main/java/com/cloudbees/sdk/commands/PluginDeleteCommand.java
@@ -5,6 +5,7 @@ import com.cloudbees.sdk.GAV;
 import com.cloudbees.sdk.cli.BeesCommand;
 import com.cloudbees.sdk.cli.CLICommand;
 import com.cloudbees.sdk.cli.CommandService;
+import com.cloudbees.sdk.utils.Helper;
 
 import javax.inject.Inject;
 
@@ -14,8 +15,14 @@ import javax.inject.Inject;
 @CLICommand("sdk:plugin:list")
 @BeesCommand(group="SDK", description = "List CLI plugins")
 public class PluginDeleteCommand extends Command {
+    private Boolean force;
+
     @Inject
     CommandService commandService;
+
+    public void setForce(Boolean force) {
+        this.force = force;
+    }
 
     public PluginDeleteCommand() {
         setArgumentExpected(1);
@@ -27,9 +34,22 @@ public class PluginDeleteCommand extends Command {
     }
 
     @Override
+    protected boolean preParseCommandLine() {
+        // add the Options
+        addOption( "f", "force", false, "force update to newest version without prompting" );
+
+        return true;
+    }
+
+    @Override
     protected boolean execute() throws Exception {
-        CommandServiceImpl service = (CommandServiceImpl) commandService;
         String name = getParameters().get(0);
+        if (force == null || !force.booleanValue()) {
+            if (!Helper.promptMatches("Are you sure you want to delete this plugin [" + name + "]: (y/n) ", "[yY].*")) {
+                return true;
+            }
+        }
+        CommandServiceImpl service = (CommandServiceImpl) commandService;
         GAV gav = service.deletePlugin(name);
         if (gav != null) {
             System.out.println("Plugin deleted: " + gav);
